Created Error Handler

+++ Restik/Restik/Views/AdminWindow.xaml.cs
@@ -44,12 +44,28 @@ namespace Restik.Views
 
         private void Button_Click_2(object sender, RoutedEventArgs e)
         {
-            FuncHelper.DeleteItem(DbManager.DeleteUser, ViewHelper.FillUsersComboBox, UsersComboBox, "Вы успешно удалили пользователя");
+            var errorMessage = ErrorHandler.GetComboBoxErrorMessage(UsersComboBox.Text, "пользователями");
+
+            if (errorMessage == null) {
+                FuncHelper.DeleteItem(DbManager.DeleteUser, ViewHelper.FillUsersComboBox, UsersComboBox, "Вы успешно удалили пользователя");
+            } else
+            {
+                ViewHelper.ShowMessage(errorMessage);
+            }
         }
 
         private void Button_Click_3(object sender, RoutedEventArgs e)
         {
-            ViewHelper.WindowsInteract(this, new UpdateUserWindow(UsersComboBox.Text));
+            var errorMessage = ErrorHandler.GetComboBoxErrorMessage(UsersComboBox.Text, "пользователями");
+
+            if (errorMessage == null)
+            {
+                ViewHelper.WindowsInteract(this, new UpdateUserWindow(UsersComboBox.Text));
+            }
+            else
+            {
+                ViewHelper.ShowMessage(errorMessage);
+            }
         }
 
         private void Button_Click_4(object sender, RoutedEventArgs e)
@@ -59,12 +75,30 @@ namespace Restik.Views
 
         private void Button_Click_5(object sender, RoutedEventArgs e)
         {
-            FuncHelper.DeleteItem(DbManager.DeleteHall, ViewHelper.FillHallsComboBox, HallsComboBox, "Вы успешно удалили зал");
+            var errorMessage = ErrorHandler.GetComboBoxErrorMessage(HallsComboBox.Text, "залами");
+
+            if (errorMessage == null)
+            {
+                FuncHelper.DeleteItem(DbManager.DeleteHall, ViewHelper.FillHallsComboBox, HallsComboBox, "Вы успешно удалили зал");
+            }
+            else
+            {
+                ViewHelper.ShowMessage(errorMessage);
+            }
         }
 
         private void Button_Click_6(object sender, RoutedEventArgs e)
         {
-            ViewHelper.WindowsInteract(this, new UpdateHallWindow(HallsComboBox.Text));
+            var errorMessage = ErrorHandler.GetComboBoxErrorMessage(HallsComboBox.Text, "залами");
+
+            if (errorMessage == null)
+            {
+                ViewHelper.WindowsInteract(this, new UpdateHallWindow(HallsComboBox.Text));
+            }
+            else
+            {
+                ViewHelper.ShowMessage(errorMessage);
+            }
         }
 
         private void Button_Click_7(object sender, RoutedEventArgs e)
@@ -74,12 +108,30 @@ namespace Restik.Views
 
         private void Button_Click_8(object sender, RoutedEventArgs e)
         {
-            ViewHelper.WindowsInteract(this, new UpdateTableWindow(TablesComboBox.Text));
+            var errorMessage = ErrorHandler.GetComboBoxErrorMessage(TablesComboBox.Text, "столами");
+
+            if (errorMessage == null)
+            {
+                ViewHelper.WindowsInteract(this, new UpdateTableWindow(TablesComboBox.Text));
+            }
+            else
+            {
+                ViewHelper.ShowMessage(errorMessage);
+            }
         }
 
         private void Button_Click_9(object sender, RoutedEventArgs e)
         {
-            FuncHelper.DeleteItem(DbManager.DeleteTable, ViewHelper.FillTablesComboBox, TablesComboBox, "Вы успешно удалили стол");
+            var errorMessage = ErrorHandler.GetComboBoxErrorMessage(TablesComboBox.Text, "столами");
+
+            if (errorMessage == null)
+            {
+                FuncHelper.DeleteItem(DbManager.DeleteTable, ViewHelper.FillTablesComboBox, TablesComboBox, "Вы успешно удалили стол");
+            }
+            else
+            {
+                ViewHelper.ShowMessage(errorMessage);
+            }
         }
 
         private void Button_Click_10(object sender, RoutedEventArgs e)
@@ -89,12 +141,30 @@ namespace Restik.Views
 
         private void Button_Click_11(object sender, RoutedEventArgs e)
         {
-            ViewHelper.WindowsInteract(this, new UpdatePlaceWindow(PlacesComboBox.Text));
+            var errorMessage = ErrorHandler.GetComboBoxErrorMessage(PlacesComboBox.Text, "местами");
+
+            if (errorMessage == null)
+            {
+                ViewHelper.WindowsInteract(this, new UpdatePlaceWindow(PlacesComboBox.Text));
+            }
+            else
+            {
+                ViewHelper.ShowMessage(errorMessage);
+            }
         }
 
         private void Button_Click_12(object sender, RoutedEventArgs e)
         {
-            FuncHelper.DeleteItem(DbManager.DeletePlace, ViewHelper.FillPlacesComboBox, PlacesComboBox, "Вы успешно удалили место");
+            var errorMessage = ErrorHandler.GetComboBoxErrorMessage(PlacesComboBox.Text, "местами");
+
+            if (errorMessage == null)
+            {
+                FuncHelper.DeleteItem(DbManager.DeletePlace, ViewHelper.FillPlacesComboBox, PlacesComboBox, "Вы успешно удалили место");
+            }
+            else
+            {
+                ViewHelper.ShowMessage(errorMessage);
+            }
         }
 
         private void Button_Click_13(object sender, RoutedEventArgs e)
@@ -104,12 +174,30 @@ namespace Restik.Views
 
         private void Button_Click_14(object sender, RoutedEventArgs e)
         {
-            ViewHelper.WindowsInteract(this, new UpdateEventWindow(EventsComboBox.Text));
+            var errorMessage = ErrorHandler.GetComboBoxErrorMessage(EventsComboBox.Text, "событиями");
+
+            if (errorMessage == null)
+            {
+                ViewHelper.WindowsInteract(this, new UpdateEventWindow(EventsComboBox.Text));
+            }
+            else
+            {
+                ViewHelper.ShowMessage(errorMessage);
+            }
         }
 
         private void Button_Click_15(object sender, RoutedEventArgs e)
         {
-            FuncHelper.DeleteItem(DbManager.DeleteEvent, ViewHelper.FillEventsComboBox, EventsComboBox, "Вы успешно удалили событие");
+            var errorMessage = ErrorHandler.GetComboBoxErrorMessage(EventsComboBox.Text, "событиями");
+
+            if (errorMessage == null)
+            {
+                FuncHelper.DeleteItem(DbManager.DeleteEvent, ViewHelper.FillEventsComboBox, EventsComboBox, "Вы успешно удалили событие");
+            }
+            else
+            {
+                ViewHelper.ShowMessage(errorMessage);
+            }
         }
 
         private void Button_Click_16(object sender, RoutedEventArgs e)
@@ -119,12 +207,30 @@ namespace Restik.Views
 
         private void Button_Click_17(object sender, RoutedEventArgs e)
         {
-            ViewHelper.WindowsInteract(this, new UpdateCuisineWindow(CuisinesComboBox.Text));
+            var errorMessage = ErrorHandler.GetComboBoxErrorMessage(CuisinesComboBox.Text, "кухнями");
+
+            if (errorMessage == null)
+            {
+                ViewHelper.WindowsInteract(this, new UpdateCuisineWindow(CuisinesComboBox.Text));
+            }
+            else
+            {
+                ViewHelper.ShowMessage(errorMessage);
+            }
         }
 
         private void Button_Click_18(object sender, RoutedEventArgs e)
         {
-            FuncHelper.DeleteItem(DbManager.DeleteCuisine, ViewHelper.FillCuisinesComboBox, CuisinesComboBox, "Вы успешно удалили кухню");
+            var errorMessage = ErrorHandler.GetComboBoxErrorMessage(CuisinesComboBox.Text, "кухнями");
+
+            if (errorMessage == null)
+            {
+                FuncHelper.DeleteItem(DbManager.DeleteCuisine, ViewHelper.FillCuisinesComboBox, CuisinesComboBox, "Вы успешно удалили кухню");
+            }
+            else
+            {
+                ViewHelper.ShowMessage(errorMessage);
+            }
         }
 
         private void Button_Click_19(object sender, RoutedEventArgs e)
@@ -134,12 +240,30 @@ namespace Restik.Views
 
         private void Button_Click_20(object sender, RoutedEventArgs e)
         {
-            ViewHelper.WindowsInteract(this, new UpdateDIshWindow(DishesComboBox.Text));
+            var errorMessage = ErrorHandler.GetComboBoxErrorMessage(DishesComboBox.Text, "блюдами");
+
+            if (errorMessage == null)
+            {
+                ViewHelper.WindowsInteract(this, new UpdateDIshWindow(DishesComboBox.Text));
+            }
+            else
+            {
+                ViewHelper.ShowMessage(errorMessage);
+            }
         }
 
         private void Button_Click_21(object sender, RoutedEventArgs e)
         {
-            FuncHelper.DeleteItem(DbManager.DeleteDish, ViewHelper.FillDishesComboBox, DishesComboBox, "Вы успешно удалили блюдо");
+            var errorMessage = ErrorHandler.GetComboBoxErrorMessage(DishesComboBox.Text, "блюдами");
+
+            if (errorMessage == null)
+            {
+                FuncHelper.DeleteItem(DbManager.DeleteDish, ViewHelper.FillDishesComboBox, DishesComboBox, "Вы успешно удалили блюдо");
+            }
+            else
+            {
+                ViewHelper.ShowMessage(errorMessage);
+            }
         }
     }
 }
